#1043 P12: relocate residual show cases — server_show.go now under 2,000 LOC
Phase 12 of #1043 — final phase. Closes the modularity-discipline
audit that started at 4,072 LOC by extracting seven residual ShowText
case bodies into a new sibling file `server_show_security_text.go`:

- `ipsec-statistics`         → `showIPsecStatistics(cfg, buf) error`
- `tunnels`                  → `showTunnels(buf)`
- `rpm`                      → `showRPM(buf)`
- `security-log`             → `showSecurityLog(filter, buf)`
- `security-alarms` /
  `security-alarms-detail`   → `showSecurityAlarms(cfg, topic, buf)`
- `schedulers`               → `showSchedulers(cfg, buf)`
- `applications`             → `showApplications(cfg, buf)`

Methodology matches Phases 1-11: semantic relocation, no behavior
change. `if … { … } else { … }` flattened into early-return form.

`showIPsecStatistics` returns `error` (the original case had a
`return nil, status.Errorf` path); the dispatcher rewraps via
`if err := …; err != nil { return nil, err }`.

`showSecurityLog` and `showSecurityAlarms` take their gRPC-request
inputs (`filter` and `topic` respectively) as parameters so the
bodies no longer reference the `req` struct directly.

| Audit progress              | LOC         |
|-----------------------------|-------------|
| Pre-#1043 server_show.go    | 4,072       |
| After Phase 1 (#1148)       | 3,945       |
| After Phase 6 (#1155)       | 2,949       |
| After Phase 11 (#1160)      | 2,202       |
| After Phase 12 (this PR)    | **1,956**   |
| Modularity threshold        | 2,000       |

server_show_security_text.go: new, 334 LOC.

All 880+ Go tests pass.
